requirements.txt
Pillow
black
click
flake8
isort
more_itertools
pre-commit
pytest
rubik-cube